Switch Mock interface to *MockClient concrete type
