+++ driver.go
@@ -13,7 +13,7 @@ var pool *mockDriver
 
 func init() {
 	pool = &mockDriver{
-		clients: make(map[string]*kivikmock),
+		clients: make(map[string]*MockClient),
 	}
 	kivik.Register("kivikmock", pool)
 }
@@ -21,7 +21,7 @@ func init() {
 type mockDriver struct {
 	sync.Mutex
 	counter int
-	clients map[string]*kivikmock
+	clients map[string]*MockClient
 }
 
 var _ driver.Driver = &mockDriver{}
@@ -35,16 +35,16 @@ func (d *mockDriver) NewClient(dsn string) (driver.Client, error) {
 		return nil, errors.New("kivikmock: no available connection found")
 	}
 	c.opened++
-	return c, nil
+	return &driverClient{MockClient: c}, nil
 }
 
 // New creates a kivik client connection and a mock to manage expectations.
-func New() (*kivik.Client, Mock, error) {
+func New() (*kivik.Client, *MockClient, error) {
 	pool.Lock()
 	dsn := fmt.Sprintf("kivikmock_%d", pool.counter)
 	pool.counter++
 
-	kmock := &kivikmock{dsn: dsn, drv: pool, ordered: true}
+	kmock := &MockClient{dsn: dsn, drv: pool, ordered: true}
 	pool.clients[dsn] = kmock
 	pool.Unlock()
